added workaround for issue on some Windows machines
The cross product produced the error "3613 - The native shader compilation failed." when running on a Windows device with Firefox. This got reported via Twitter, and I could reproduce it on my Windows 10 computer. I will report this to Adobe; but for now, calculating the cross product manually does the trick.