fixed wrong specular lighting results on iOS
On iOS, the vector normalization produced a wrong result.
My guess: the "square" operation that's part of the normalization process creates a too high value.
Scaling the vector down to 10% before normalizing fixes the issue.

The downside: in "baseline constrained" profile, the fragment program is now too long to work with both an ambient and a point/directional light.